requirements.txt
# Lumina-Red AI Red Teaming Platform
openai>=1.12.0
anthropic>=0.18.0
httpx>=0.27.0
websockets>=12.0
rich>=13.7.0
python-dotenv>=1.0.0
typer>=0.12.0
